Ready for its first release.
There were a number of major design changes since the last commit:

(Preamble)
On December 17th, 2015 I sat down with all of the Image Inspector's
users (all five of them at DNVGL). The most popular complaint from
them was that the Image Inspector was too hard to use. Many were
annoyed by how much work it took to just zoom-in on an image. It was
just a click, sometimes two clicks, but that was enough to put them
off from using the Image Inspector as their main viewer. Another
popular complaint was that the concept of 'layers' - having to combine
them, re-order them, think about them - was too complicated. I will
finish off with a quote: "I want it to be like the Windows Photo
Viewer, but with the screen ruler".

 - I got rid of the transform extension. Its now built into SIV. Now,
   you can zoom in/out on an image by just scrolling up and down on
   the image layer. Just like with any other image viewer.
 - I got rid of the layers extension. Now, a user just makes their
   measurement, then types in Ctrl + S, and all layers in the viewer
   get combined and exported directly onto the file system.
 - I added a Help menu item with two commands:
   1. Report a Bug - this brings up a Google Form that gathers
   information on the issues a user is experiencing.
   2. Report a Pain Point - this brings up a Google Form that gathers
   information on the parts of SIV that annoy them, or the parts that
   they thing could be enhanced in some way.