I went a little crazy on this one.
During regression testing I noticed a bug where a SIV window wouldn't
open when logged in as a 'guest', 'developer', or 'no-connection'. It
was the weirdest bug, and I still don't understand it. I ended up
modularizing and re-writing most of app.js just to get a better
understanding of what was going on. And to be honest, I still don't
understand what happened. I "solved it" by trial and error. I noticed
that two things would make the bug disappear and cause the window to
appear as expected:

 1. When I added a console.log statement somewhere in the handleInput
 module. I know, weird right? This made me think that there was an
 error being swallowed by one of the Promises, but I checked for
 errors using the methods described here:
 http://jamesknelson.com/are-es6-promises-swallowing-your-errors/, and
 nothing came up.

 2. The only difference between being logged in as a 'guest',
 'developer', 'no-connection', and being logged in as a numbered user,
 is that SIV would check for a yubikey every 3 seconds. I noticed that
 if I repeated any function - even an empty one (() => {}) - that the
 bug would suddenly disappear and the SIV window would appear as
 expected. WTF. I ended up 'solving' the bug by creating a setInterval
 for any logged-in user.