Add openedExtensions key to the siv store
The openedExtensions key will store an array of extension path/package
names that a user opened. An open extension has a layer in the viewer,
and a controls component in the sidebar. The order of the array stored
at the openedExtensions key determines the order of the layers in the
viewer and the controls in the sidebar.