Add webpack support (#16)
* Rename availableExtensions to installedExtensions

* Add openedExtensions key to the siv store

The openedExtensions key will store an array of extension path/package
names that a user opened. An open extension has a layer in the viewer,
and a controls component in the sidebar. The order of the array stored
at the openedExtensions key determines the order of the layers in the
viewer and the controls in the sidebar.

* Upgrade electron to v1.2.7

* Create Redux store for extensions

* Add webpack support

React is now passed down to the extensions. This ensures that only one
react instance runs at a time. A new global variable 'dynamicRequire' is
now globally available in siv's renderer process. 'dynamicRequire' was
made global so it would be ignored by webpack.